refactor: move ecstore cluster control plane module
