+++ docs/architecture/migration-progress.md
@@ -5,18 +5,18 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 ## Current Context
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
-- Branch: `overtrue/arch-ecstore-core-runtime-layout`
+- Branch: `overtrue/arch-ecstore-cluster-control-plane-layout`
 - Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117/API-118/API-119/API-120/API-121/API-122/API-123/API-124/API-125/API-126/API-127/API-128/API-129/API-130/API-131/API-132/API-133/API-134/API-135/API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146/API-147/API-148/API-149/API-150/API-151/API-152/API-153/API-154/API-155/API-156/API-157/API-158/API-159/API-160/API-161/API-162/API-163/API-164/API-165/API-166/API-167/API-168/API-169/API-170/API-171/API-172/API-173/API-174/API-175/API-176/API-177/API-178/API-179/API-180/API-181/API-182/API-183/API-184/API-185/API-186/API-187/API-188/API-189/API-190/API-191/API-192/API-193/API-194/API-195/API-196/API-197/API-198/API-199/API-200/API-201/API-202/API-203/API-204/API-205/API-206/API-207/API-208/API-209/API-210/API-211/API-212/API-213/API-214/API-215/API-216/API-217/API-218/API-219/API-220/API-221/API-222/API-223/API-224/API-225/API-226/API-227/API-228/API-229/API-230/API-231/API-232/API-233/API-234/API-235/API-236/API-237/API-238/API-239/API-240/API-241/API-242/API-243/API-244/API-245/API-246/API-247/API-248/API-249/API-250/API-251/API-252/API-253/API-254/CTX-002`.
 - Current baseline also includes API-255 from PR #3923, API-256 from PR
   #3925, and CFG-009 from PR #3927.
-- Current phase PR: E-027/E-CORE-001 ECStore core runtime owner layout.
-- Based on: E-026 branch while PR #3930 is pending; rebase onto `origin/main`
-  after E-017 through E-026 merge before opening this PR.
+- Current phase PR: E-028/E-CLUSTER-001 ECStore cluster control-plane owner layout.
+- Based on: E-027 branch while PR #3930 is pending; rebase onto `origin/main`
+  after E-017 through E-027 merge before opening this PR.
 - PR type for this branch: `pure-move`.
-- Runtime behavior changes: none expected for E-027; production changes are
-  module file moves that keep the existing `pools` and `sets` module paths.
-- Rust code changes: move ECStore core runtime and test harness modules under
-  the `core` owner directory without function-body changes.
+- Runtime behavior changes: none expected for E-028; production changes are a
+  module file move that keeps the existing `cluster` facade paths.
+- Rust code changes: move ECStore cluster control-plane implementation under
+  the `cluster/control_plane` owner module without function-body changes.
 - CI/script changes: reject restoring ECStore root `store.rs`, `set_disk.rs`,
   `store_list_objects.rs`, `store_utils.rs`, `store_init.rs`,
   `disks_layout.rs`, `endpoints.rs`, `storage_api_contracts.rs`,
@@ -70,7 +70,8 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
   root runtime global modules after E-024, and reject restoring ECStore root
   I/O support modules after E-025, and reject restoring ECStore root error and
   rebalance facade modules after E-026, and reject restoring ECStore root core
-  runtime/test modules after E-027.
+  runtime/test modules after E-027, and keep the ECStore cluster root module as
+  a re-export-only owner facade after E-028.
 
 ## Phase 0 Tasks
 
@@ -4084,6 +4085,21 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
     formatting, diff hygiene, Rust risk scan, branch freshness check,
     pre-commit quality gate, and three-expert review.
 
+- [x] `E-028/E-CLUSTER-001` Move ECStore cluster control-plane implementation under owner.
+  - Do: move `crates/ecstore/src/cluster/mod.rs` implementation into
+    `crates/ecstore/src/cluster/control_plane.rs` and keep
+    `cluster/mod.rs` as a re-export-only facade.
+  - Acceptance: `rustfs_ecstore::api::cluster::*` and existing
+    `crate::cluster::*` consumers keep the same symbols, and migration rules
+    reject restoring control-plane implementation declarations to the cluster
+    root module.
+  - Must preserve: topology snapshots, membership grouping, pool-state
+    reporting, local-node storage projection, peer-health unknown status,
+    read-only facade behavior, and all cluster tests.
+  - Verification: focused ECStore compile/tests, migration and layer guards,
+    formatting, diff hygiene, Rust risk scan, branch freshness check,
+    pre-commit quality gate, and three-expert review.
+
 - [x] `API-129` Route RustFS internal ECStore consumers through owner boundary.
   - Do: expose crate-local ECStore facade module aliases from
     `rustfs/src/storage/mod.rs` and migrate RustFS startup, server, capacity,
@@ -5938,6 +5954,9 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 
 | Expert | Status | Notes |
 |---|---|---|
+| Quality/architecture | pass | E-028 moves the ECStore cluster control-plane implementation into an owner submodule and leaves the root module as a facade. |
+| Migration preservation | pass | `crate::cluster::*` and `rustfs_ecstore::api::cluster::*` stay stable through root re-exports; the moved implementation is a 100% rename. |
+| Testing/verification | pass | ECStore cluster tests, ECStore full tests, architecture/layer guards, formatting, diff hygiene, diff-added Rust risk scan, and pre-commit passed. |
 | Quality/architecture | pass | E-027 moves ECStore pools, sets, and root core test harness modules under the core owner directory and extends the root-file guard. |
 | Migration preservation | pass | `crate::pools`, `crate::sets`, `pools_test`, and `store_test` stay stable through path declarations; moved files are 100% renames. |
 | Testing/verification | pass | ECStore full tests, architecture/layer guards, formatting, diff hygiene, diff-added Rust risk scan, and pre-commit passed. |
